api/requirements.txt
fastAPI[standard]
pydantic
jose
passlib
uvicorn
databases[asyncpg]
SQLAlchemy
python-jose
bcrypt
python-dotenv
requests
psycopg2